w3c-reports/12092014/gtk/runned-tests/html/semantics/forms/the-input-element/hidden-expected.txt

Hidden input element


PASS Value returns the current value for hidden 
PASS Setting value changes the current value for hidden 
PASS files attribute must return null for hidden 
PASS valueAsDate attribute must return null for hidden 
PASS valueAsNumber attribute must return NaN for hidden 
FAIL list attribute must return null for hidden assert_equals: expected (object) null but got (undefined) undefined(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/html/semantics/forms/the-input-element/hidden.html:53:22
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/the-input-element/hidden.html:51:9)
PASS stepDown does not apply for hidden 
PASS stepUp does not apply for hidden 
PASS input type=hidden is barred from constraint validation 
